Schema store rewrite.
- The new schema store uses the property store to represent the schema rules as maps of properties.
- This allows us to let the index providers store nearly arbitrary configurations about each index, by simply adding more non-conflicting properties to the schema rules.
- This hasn't been done yet. So far we are only about laying the ground work and providing the infrastructure to do this.
- The key aspect of this, is that the complete schema rule map will also be serialised into the transaction log.
- This means that index configurations, at the end of this journey, will also be replicated in a clustered environment.

+++ community/record-storage-engine/src/test/java/org/neo4j/kernel/impl/store/PropertyStoreConsistentReadTest.java
@@ -79,7 +79,7 @@ class PropertyStoreConsistentReadTest extends RecordStoreConsistentReadTest<Prop
     }
 
     @Override
-    protected PropertyRecord getHeavy( PropertyStore store, int id )
+    protected PropertyRecord getHeavy( PropertyStore store, long id )
     {
         PropertyRecord record = super.getHeavy( store, id );
         ensureHeavy( store, record );
